Refactor cv code; #1035
